Refactor: Split sensor.py into modular structure
BREAKING CHANGE: sensor.py restructure (functional behavior unchanged)

Split large sensor.py file (1102 lines) into focused modules:
- sensor_modules/__init__.py: Module exports
- sensor_modules/base.py: Constants and helper functions
- sensor_modules/generic.py: VioletSensor and VioletStatusSensor
- sensor_modules/specialized.py: ErrorCode, Dosing, FlowRate sensors
- sensor_modules/monitoring.py: 6 monitoring sensors (SystemHealth, Latency, etc.)
- sensor.py: Reduced to 270 lines (setup functions only, -75%)

Changes:
- Extracted 10 sensor classes into logical groupings
- Created _build_sensor_description() helper in base.py
- Fixed import dependencies between modules
- Added proper exports in __init__.py
- Maintained backward compatibility - all sensors work identically

Testing:
- 376 sensors created successfully
- No import errors
- All sensor types functional
- Setup completed successfully in Docker

Related: Previous config_flow.py refactoring follows same pattern
