+++ custom_components/violet_pool_controller/sensor.py
@@ -1,26 +1,17 @@
 """Sensor integration for the Violet Pool Controller.
 
-This module defines the sensor entities for the integration, including
-generic sensors, status sensors, and specialized sensors for error codes and
-flow rates. It also handles the logic for creating and configuring these
-sensors based on the user's settings.
+This module defines the sensor setup and entity creation logic for the integration.
+Sensor classes are split into submodules for better organization.
 """
 
 from __future__ import annotations
 
 import logging
-from datetime import datetime, timezone
 from typing import Any
 
-from homeassistant.components.sensor import (
-    SensorDeviceClass,
-    SensorEntity,
-    SensorEntityDescription,
-    SensorStateClass,
-)
+from homeassistant.components.sensor import SensorEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import (
@@ -36,804 +27,32 @@ from .const import (
     TEMP_SENSORS,
     UNIT_MAP,
     WATER_CHEM_SENSORS,
-    VioletState,
 )
 from .device import VioletPoolDataUpdateCoordinator
-from .entity import VioletPoolControllerEntity
 from .error_codes import get_error_info
 
-# Type checking for coordinator
-# VioletPoolDataUpdateCoordinator is already imported above
+# Import sensor classes from submodules
+from .sensor_modules import (
+    VioletAPIRequestRateSensor,
+    VioletAverageLatencySensor,
+    VioletConnectionLatencySensor,
+    VioletDosingStateSensor,
+    VioletErrorCodeSensor,
+    VioletFlowRateSensor,
+    VioletLastEventAgeSensor,
+    VioletRecoverySuccessRateSensor,
+    VioletSensor,
+    VioletStatusSensor,
+    VioletSystemHealthSensor,
+    _build_sensor_description,
+    should_skip_sensor,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
-# --- Sensor Categorization ---
-
-_TIMESTAMP_SUFFIXES = (
-    "_LAST_ON",
-    "_LAST_OFF",
-    "_LAST_AUTO_RUN",
-    "_LAST_MANUAL_RUN",
-    "_LAST_CAN_RESET",
-    "_TIMESTAMP",
-)
-_TIMESTAMP_KEYS = {"CURRENT_TIME_UNIX"} | {
-    key
-    for key in UNIT_MAP
-    if any(key.upper().endswith(suffix) for suffix in _TIMESTAMP_SUFFIXES)
-}
-
-_BOOLEAN_VALUE_KEYS = {
-    "OVERFLOW_REFILL_STATE",
-    "OVERFLOW_DRYRUN_STATE",
-    "OVERFLOW_OVERFILL_STATE",
-    "BACKWASH_OMNI_MOVING",
-    "BACKWASH_DELAY_RUNNING",
-    "BATHING_AI_SURVEILLANCE_STATE",
-    "BATHING_AI_PUMP_STATE",
-} | {f"onewire{i}_state" for i in range(1, 13)}
-
-_RUNTIME_KEYS = (
-    {
-        "ECO_RUNTIME",
-        "LIGHT_RUNTIME",
-        "PUMP_RUNTIME",
-        "BACKWASH_RUNTIME",
-        "BACKWASHRINSE_RUNTIME",
-        "SOLAR_RUNTIME",
-        "HEATER_RUNTIME",
-        "DOS_1_CL_RUNTIME",
-        "DOS_2_ELO_RUNTIME",
-        "DOS_3_ELO_REV_RUNTIME",
-        "DOS_4_PHM_RUNTIME",
-        "DOS_5_PHP_RUNTIME",
-        "DOS_6_FLOC_RUNTIME",
-        "REFILL_RUNTIME",
-        "REFILL_TIMEOUT",
-        "RUNTIME",
-        "POSTRUN_TIME",
-    }
-    | {f"EXT{i}_{j}_RUNTIME" for i in (1, 2) for j in range(1, 9)}
-    | {f"OMNI_DC{i}_RUNTIME" for i in range(6)}
-    | {f"PUMP_RPM_{i}_RUNTIME" for i in range(4)}
-)
-
-_TIME_FORMAT_KEYS = (
-    {
-        "HEATER_POSTRUN_TIME",
-        "SOLAR_POSTRUN_TIME",
-        "CPU_UPTIME",
-        "DEVICE_UPTIME",
-    }
-    | {f"PUMP_RPM_{i}_LAST_ON" for i in range(4)}
-    | {f"PUMP_RPM_{i}_LAST_OFF" for i in range(4)}
-)
-
-_TEXT_VALUE_KEYS = {
-    "DOS_1_CL_STATE",
-    "DOS_2_ELO_STATE",
-    "DOS_4_PHM_STATE",
-    "DOS_5_PHP_STATE",
-    "DOS_6_FLOC_STATE",
-    "HEATERSTATE",
-    "SOLARSTATE",
-    "PUMPSTATE",
-    "BACKWASHSTATE",
-    # STATUS_SENSORS - these return text values like "Automatik (Bereit)", "Ein", "Aus"
-    "BACKWASH",
-    "PUMP",
-    "HEATER",
-    "SOLAR",
-    "LIGHT",
-    "PVSURPLUS",
-    # Other STATE sensors
-    "OMNI_STATE",
-    "BACKWASH_OMNI_STATE",
-    "SOLAR_STATE",
-    "HEATER_STATE",
-    "PUMP_STATE",
-    "FILTER_STATE",
-    "OMNI_MODE",
-    "FILTER_MODE",
-    "SOLAR_MODE",
-    "HEATER_MODE",
-    "DISPLAY_MODE",
-    "OPERATING_MODE",
-    "MAINTENANCE_MODE",
-    "ERROR_CODE",
-    "LAST_ERROR",
-    "VERSION_CODE",
-    "CHECKSUM",
-    "RULE_RESULT",
-    "LAST_MOVING_DIRECTION",
-    "COVER_STATE",
-    "COVER_DIRECTION",
-    "SW_VERSION",
-    "SW_VERSION_CARRIER",
-    "HW_VERSION_CARRIER",
-    "FW",
-    "fw",
-    "VERSION",
-    "VERSION_INFO",
-    "HARDWARE_VERSION",
-    "CPU_GOV",
-    "HW_SERIAL_CARRIER",
-    "SERIAL_NUMBER",
-    "MAC_ADDRESS",
-    "IP_ADDRESS",
-    "DOS_1_CL_REMAINING_RANGE",
-    "DOS_4_PHM_REMAINING_RANGE",
-    "DOS_5_PHP_REMAINING_RANGE",
-    "DOS_6_FLOC_REMAINING_RANGE",
-    "BACKWASH_OMNI_MOVING",
-    "BACKWASH_DELAY_RUNNING",
-    "BACKWASH_STATE",
-    "REFILL_STATE",
-    "BATHING_AI_SURVEILLANCE_STATE",
-    "BATHING_AI_PUMP_STATE",
-    "OVERFLOW_REFILL_STATE",
-    "OVERFLOW_DRYRUN_STATE",
-    "OVERFLOW_OVERFILL_STATE",
-    "time",
-    "TIME",
-    "CURRENT_TIME",
-    "date",  # Date value, not numeric
-    "LOAD_AVG",
-    "pump_rs485_pwr",
-    "SYSTEM_carrier_alive_count",
-    "SYSTEM_carrier_alive_faultcount",
-    "SYSTEM_dosagemodule_alive_count",
-    "SYSTEM_dosagemodule_alive_faultcount",
-    "SYSTEM_ext1module_alive_count",
-    "SYSTEM_ext1module_alive_faultcount",
-    # Cover contact sensors - these return string values ("RELEASED"/"TRIGGERED")
-    "CLOSE_CONTACT",
-    "OPEN_CONTACT",
-    "ERROR_CONTACT",
-}
-
-_ALL_TEXT_SENSORS = (
-    _TEXT_VALUE_KEYS | _RUNTIME_KEYS | _BOOLEAN_VALUE_KEYS | _TIME_FORMAT_KEYS
-)
-
-_NON_TEMPERATURE_ONEWIRE_KEYS = {
-    f"onewire{i}_{suffix}"
-    for i in range(1, 13)
-    for suffix in ("rcode", "romcode", "state")
-}
-_FLOW_RATE_SOURCE_KEYS = {"ADC3_value", "IMP2_value"}
+# Additional constants for sensor classification
 _ERROR_CODE_KEYS = {"LAST_ERROR_CODE", "ERROR_CODE", "LAST_ERROR"}
-
-
-class VioletSensor(VioletPoolControllerEntity, SensorEntity):
-    """Represents a generic Violet Pool Controller sensor."""
-
-    entity_description: SensorEntityDescription
-    coordinator: VioletPoolDataUpdateCoordinator
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-        description: SensorEntityDescription,
-    ) -> None:
-        """Initializes the sensor.
-
-        Args:
-            coordinator: The data update coordinator.
-            config_entry: The configuration entry.
-            description: The entity description for the sensor.
-        """
-        super().__init__(coordinator, config_entry, description)
-        self._logger = logging.getLogger(f"{DOMAIN}.sensor.{description.key}")
-        _LOGGER.debug(
-            "Sensor initialized: %s (Key: %s, Class: %s)",
-            description.name or description.translation_key,
-            description.key,
-            description.device_class,
-        )
-
-    @property
-    def state_class(self) -> SensorStateClass | None:
-        """Override state_class for contact sensors to prevent numeric conversion errors.
-
-        Contact sensors return string values ('RELEASED', 'TRIGGERED') but may have been
-        incorrectly created with state_class='measurement'. This property override
-        ensures they always return None, preventing Home Assistant from attempting
-        numeric conversion.
-        """
-        # Force state_class to None for contact sensors
-        if "contact" in self.entity_description.key.lower():
-            _LOGGER.debug(
-                "Overriding state_class to None for contact sensor: %s",
-                self.entity_description.key,
-            )
-            return None
-        return self.entity_description.state_class
-
-    @property
-    def native_value(self) -> str | int | float | datetime | None:
-        """Returns the native value of the sensor, formatted for Home Assistant."""
-        if self.coordinator.data is None:
-            return None
-
-        key = self.entity_description.key
-        raw_value = self.coordinator.data.get(key)
-
-        if raw_value is None:
-            return None
-
-        # Check if key indicates a timestamp sensor (by suffix or membership in _TIMESTAMP_KEYS)
-        is_timestamp_key = key in _TIMESTAMP_KEYS or any(
-            key.upper().endswith(suffix) for suffix in _TIMESTAMP_SUFFIXES
-        )
-
-        if is_timestamp_key and key not in _TIME_FORMAT_KEYS:
-            try:
-                timestamp = float(raw_value)
-                # Handle milliseconds: if timestamp > 10000000000, it's in milliseconds
-                # (10000000000 ms = September 2001, in seconds since 1970)
-                if timestamp > 10000000000:
-                    timestamp = timestamp / 1000
-                return datetime.fromtimestamp(timestamp, tz=timezone.utc)
-            except (ValueError, TypeError) as err:
-                self._logger.warning(
-                    "Timestamp conversion failed for %s with value '%s': %s",
-                    key,
-                    raw_value,
-                    err,
-                )
-                return None
-
-        if key in _ALL_TEXT_SENSORS:
-            return str(raw_value)
-
-        try:
-            num_value = float(raw_value)
-
-            # Water chemistry values (pH, ORP, Chlorine) - 2 decimal places for precision
-            if key in {"pH_value", "orp_value", "pot_value"}:
-                return round(num_value, 2)
-
-            # Temperature sensors (all onewire, CPU temps) - 2 decimal places
-            # IMPORTANT: Exclude freezecount, faultcount - these are counters, NOT temperatures!
-            if ("temp" in key.lower() or "onewire" in key.lower()) and "freezecount" not in key.lower() and "faultcount" not in key.lower():
-                return round(num_value, 2)
-
-            # Analog sensors (ADC, IMP) - 2 decimal places for precision
-            if key.startswith(("ADC", "IMP")):
-                return round(num_value, 2)
-
-            # Percentage values - 1 decimal place
-            if key.startswith(("SYSTEM_")) or "_" in key and key.split("_")[-1] in ["PERCENT", "PERCENTAGE"]:
-                return round(num_value, 1)
-
-            # Integer values (counts, RPM, etc.) - round to integer if close to whole number
-            if num_value.is_integer():
-                return int(num_value)
-
-            # All other numeric values - 2 decimal places for consistency
-            return round(num_value, 2)
-
-        except (ValueError, TypeError):
-            # Explicitly cast to string to match return type
-            return str(raw_value)
-
-
-class VioletStatusSensor(VioletSensor):
-    """Represents a sensor for status values that use VioletState."""
-
-    def _resolve_raw_value(self) -> Any | None:
-        """Resolve the best raw value, preferring *STATE key with fallback."""
-        key = self.entity_description.key
-        state_key = f"{key}STATE"
-
-        # Prefer *STATE field (e.g., PUMPSTATE = "3|PUMP_ANTI_FREEZE")
-        raw_value = self.coordinator.data.get(state_key)
-
-        # Skip empty/useless *STATE values (e.g., SOLARSTATE = "[]")
-        if raw_value is not None and str(raw_value).strip() in ("", "[]", "{}"):
-            raw_value = None
-
-        if raw_value is None:
-            raw_value = self.coordinator.data.get(key)
-
-        return raw_value
-
-    @property
-    def native_value(self) -> str | int | float | datetime | None:
-        """Return the display string for the status."""
-        if self.coordinator.data is None:
-            return None
-
-        raw_value = self._resolve_raw_value()
-        return (
-            VioletState(raw_value, self.entity_description.key).display_mode
-            if raw_value is not None
-            else None
-        )
-
-    @property
-    def icon(self) -> str | None:
-        """Return the icon corresponding to the current status."""
-        if self.coordinator.data is None:
-            return super().icon
-
-        raw_value = self._resolve_raw_value()
-        if raw_value is None:
-            return super().icon
-        return VioletState(raw_value, self.entity_description.key).icon
-
-
-class VioletErrorCodeSensor(VioletSensor):
-    """A specialized sensor that resolves error codes to descriptive text."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-        value_key: str,
-    ) -> None:
-        """Initializes the error code sensor.
-
-        Args:
-            coordinator: The data update coordinator.
-            config_entry: The configuration entry.
-            value_key: The key in the coordinator data that holds the error code.
-        """
-        description = SensorEntityDescription(
-            key=value_key,
-            translation_key="last_error_code",
-            name="Last Error Code",
-            icon="mdi:alert-circle",
-            entity_category=EntityCategory.DIAGNOSTIC,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> str | None:
-        """Return the descriptive subject of the error code."""
-        if self.coordinator.data is None:
-            return None
-
-        code = str(self.coordinator.data.get(self.entity_description.key, "")).strip()
-        return get_error_info(code)["subject"] if code else None
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return detailed information about the error code as attributes."""
-        if self.coordinator.data is None:
-            return {}
-
-        code = str(self.coordinator.data.get(self.entity_description.key, "")).strip()
-        if not code:
-            return {}
-        info = get_error_info(code)
-        return {
-            "code": code,
-            "type": info.get("type"),
-            "severity": info.get("severity"),
-            "description": info.get("description"),
-        }
-
-
-class VioletSystemHealthSensor(VioletPoolControllerEntity, SensorEntity):
-    """Diagnostic sensor for system health percentage (0-100)."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the system health sensor.
-
-        Args:
-            coordinator: The data update coordinator.
-            config_entry: The configuration entry.
-        """
-        description = SensorEntityDescription(
-            key="system_health",
-            translation_key="system_health",
-            name="System Health",
-            icon="mdi:heart-pulse",
-            entity_category=EntityCategory.DIAGNOSTIC,
-            native_unit_of_measurement="%",
-            state_class=SensorStateClass.MEASUREMENT,
-            suggested_display_precision=0,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> float | None:
-        """Return the system health percentage."""
-        return round(self.coordinator.device.system_health, 0)
-
-
-class VioletConnectionLatencySensor(VioletPoolControllerEntity, SensorEntity):
-    """Diagnostic sensor for connection latency in milliseconds."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the connection latency sensor.
-
-        Args:
-            coordinator: The data update coordinator.
-            config_entry: The configuration entry.
-        """
-        description = SensorEntityDescription(
-            key="connection_latency",
-            translation_key="connection_latency",
-            name="Connection Latency",
-            icon="mdi:speedometer",
-            entity_category=EntityCategory.DIAGNOSTIC,
-            native_unit_of_measurement="ms",
-            state_class=SensorStateClass.MEASUREMENT,
-            device_class=SensorDeviceClass.DURATION,
-            suggested_display_precision=0,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> float | None:
-        """Return the connection latency in milliseconds."""
-        return round(self.coordinator.device.connection_latency, 0)
-
-
-class VioletLastEventAgeSensor(VioletPoolControllerEntity, SensorEntity):
-    """Diagnostic sensor for seconds since last successful update."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the last event age sensor.
-
-        Args:
-            coordinator: The data update coordinator.
-            config_entry: The configuration entry.
-        """
-        description = SensorEntityDescription(
-            key="last_event_age",
-            translation_key="last_event_age",
-            name="Last Event Age",
-            icon="mdi:clock-outline",
-            entity_category=EntityCategory.DIAGNOSTIC,
-            native_unit_of_measurement="s",
-            state_class=SensorStateClass.MEASUREMENT,
-            device_class=SensorDeviceClass.DURATION,
-            suggested_display_precision=0,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> float | None:
-        """Return seconds since last successful update."""
-        return round(self.coordinator.device.last_event_age, 0)
-
-
-class VioletAPIRequestRateSensor(VioletPoolControllerEntity, SensorEntity):
-    """Sensor for API request rate (requests per minute)."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the API request rate sensor."""
-        description = SensorEntityDescription(
-            key="api_request_rate",
-            translation_key="api_request_rate",
-            name="API Request Rate",
-            icon="mdi:network",
-            native_unit_of_measurement="req/min",
-            entity_category=EntityCategory.DIAGNOSTIC,
-            state_class=SensorStateClass.MEASUREMENT,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> float | None:
-        """Return the API request rate."""
-        return round(self.coordinator.device.api_request_rate, 1)
-
-
-class VioletAverageLatencySensor(VioletPoolControllerEntity, SensorEntity):
-    """Sensor for average connection latency."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the average latency sensor."""
-        description = SensorEntityDescription(
-            key="average_latency",
-            translation_key="average_latency",
-            name="Average Latency",
-            icon="mdi:speedometer",
-            native_unit_of_measurement="ms",
-            entity_category=EntityCategory.DIAGNOSTIC,
-            state_class=SensorStateClass.MEASUREMENT,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> float | None:
-        """Return the average connection latency."""
-        return round(self.coordinator.device.average_latency, 0)
-
-
-class VioletRecoverySuccessRateSensor(VioletPoolControllerEntity, SensorEntity):
-    """Sensor for recovery success rate."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initialize the recovery success rate sensor."""
-        description = SensorEntityDescription(
-            key="recovery_success_rate",
-            translation_key="recovery_success_rate",
-            name="Recovery Success Rate",
-            icon="mdi:restore",
-            native_unit_of_measurement="%",
-            entity_category=EntityCategory.DIAGNOSTIC,
-            state_class=SensorStateClass.MEASUREMENT,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> float | None:
-        """Return the recovery success rate."""
-        return round(self.coordinator.device.recovery_success_rate, 1)
-
-
-class VioletDosingStateSensor(VioletPoolControllerEntity, SensorEntity):
-    """Sensor for dosing system state arrays (DOS_*_STATE) and composite states."""
-
-    # German translations for common state detail codes
-    _DETAIL_DE: dict[str, str] = {
-        "PUMP_ANTI_FREEZE": "Frostschutz",
-        "BLOCKED_BY_OUTSIDE_TEMP": "Blockiert (Außentemp.)",
-        "BLOCKED_BY_TRESHOLDS": "Blockiert (Grenzwerte)",
-        "TRESHOLDS_REACHED": "Grenzwerte erreicht",
-        "BLOCKED_BY_PUMP": "Blockiert (Pumpe aus)",
-        "BLOCKED_BY_FLOW": "Blockiert (Durchfluss)",
-        "BLOCKED_BY_SOLAR": "Blockiert (Solar)",
-        "BLOCKED_BY_HEATER": "Blockiert (Heizung)",
-        "WAITING_FOR_PUMP": "Wartet auf Pumpe",
-        "WAITING_FOR_FLOW": "Wartet auf Durchfluss",
-        "DOSING": "Dosiert",
-        "DOSING_PAUSED": "Dosierung pausiert",
-        "MANUAL_DOSING": "Manuelle Dosierung",
-    }
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-        key: str,
-        name: str,
-        icon: str,
-    ) -> None:
-        """Initialize the dosing state sensor."""
-        description = SensorEntityDescription(
-            key=key,
-            name=name,
-            icon=icon,
-            entity_category=EntityCategory.DIAGNOSTIC,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    def _translate_detail(self, code: str) -> str:
-        """Translate a detail code to German, fallback to title-cased string."""
-        return self._DETAIL_DE.get(code, code.replace("_", " ").title())
-
-    @property
-    def native_value(self) -> str | None:
-        """Return the dosing state as comma-separated German string."""
-        raw_value = self.get_value(self.entity_description.key)
-
-        # Handle array values (DOS_*_STATE sensors)
-        if isinstance(raw_value, list):
-            if not raw_value:
-                return "OK"
-            return ", ".join(self._translate_detail(str(item)) for item in raw_value)
-
-        # Handle pipe-separated strings (PUMPSTATE, HEATERSTATE, SOLARSTATE)
-        if isinstance(raw_value, str) and "|" in raw_value:
-            parts = raw_value.split("|", 1)
-            if len(parts) == 2:
-                return self._translate_detail(parts[1])
-            return raw_value
-
-        # Handle string values
-        if isinstance(raw_value, str):
-            return raw_value if raw_value else "OK"
-
-        return None
-
-    @property
-    def extra_state_attributes(self) -> dict[str, Any]:
-        """Return additional attributes."""
-        raw_value = self.get_value(self.entity_description.key)
-
-        attributes: dict[str, Any] = {}
-
-        # Array-based sensors (DOS_*_STATE)
-        if isinstance(raw_value, list):
-            attributes["state_count"] = len(raw_value)
-            attributes["states_list"] = raw_value
-            attributes["has_issues"] = len(raw_value) > 0
-
-        # Pipe-separated sensors (PUMPSTATE, HEATERSTATE, SOLARSTATE)
-        elif isinstance(raw_value, str) and "|" in raw_value:
-            parts = raw_value.split("|", 1)
-            if len(parts) == 2:
-                state_num, detail = parts
-                attributes["numeric_state"] = state_num
-                attributes["detail_code"] = detail
-                attributes["detail_readable"] = self._translate_detail(detail)
-                attributes["raw_value"] = raw_value
-
-        return attributes
-
-
-class VioletFlowRateSensor(VioletPoolControllerEntity, SensorEntity):
-    """A specialized sensor for flow rate that prioritizes ADC3 over IMP2."""
-
-    def __init__(
-        self,
-        coordinator: VioletPoolDataUpdateCoordinator,
-        config_entry: ConfigEntry,
-    ) -> None:
-        """Initializes the flow rate sensor."""
-        description = SensorEntityDescription(
-            key="flow_rate_adc3_priority",
-            translation_key="flow_rate",
-            name="Flow Rate",
-            icon="mdi:pump",
-            native_unit_of_measurement="m³/h",
-            device_class=SensorDeviceClass.VOLUME_FLOW_RATE,
-            state_class=SensorStateClass.MEASUREMENT,
-        )
-        super().__init__(coordinator, config_entry, description)
-
-    @property
-    def native_value(self) -> float | None:
-        """Return the flow rate, prioritizing the ADC3 value."""
-        if self.coordinator.data is None:
-            return None
-
-        for key in ["ADC3_value", "IMP2_value"]:
-            raw_value = self.coordinator.data.get(key)
-            if raw_value is not None:
-                try:
-                    return round(float(raw_value), 2)
-                except (ValueError, TypeError):
-                    continue
-        return None
-
-    @property
-    def extra_state_attributes(self) -> dict[str, str]:
-        """Return the raw source values for debugging purposes."""
-        if self.coordinator.data is None:
-            return {"data_source": "None"}
-
-        adc3 = self.coordinator.data.get("ADC3_value", "N/A")
-        imp2 = self.coordinator.data.get("IMP2_value", "N/A")
-        source = "ADC3" if adc3 != "N/A" else "IMP2" if imp2 != "N/A" else "None"
-        return {
-            "adc3_raw_value": str(adc3),
-            "imp2_raw_value": str(imp2),
-            "data_source": source,
-        }
-
-    @property
-    def available(self) -> bool:
-        """The sensor is available if either of its data sources is present."""
-        if self.coordinator.data is None:
-            return False
-
-        return super().available and any(
-            self.coordinator.data.get(key) is not None for key in _FLOW_RATE_SOURCE_KEYS
-        )
-
-
-def _is_boolean_value(value: Any) -> bool:
-    """Checks if a value can be interpreted as a boolean."""
-    return str(value).lower().strip() in (
-        "true",
-        "false",
-        "1",
-        "0",
-        "on",
-        "off",
-        "yes",
-        "no",
-    )
-
-
-def determine_device_class(
-    key: str, unit: str | None, raw_value: Any
-) -> SensorDeviceClass | None:
-    """Determines the appropriate device class for a sensor."""
-    if key in _BOOLEAN_VALUE_KEYS or _is_boolean_value(raw_value):
-        return None
-    if key == "pH_value":
-        return SensorDeviceClass.PH
-    # Temperature device class only for actual temperature sensors (not counters)
-    if (unit == "°C" or ("temp" in key.lower() and unit is not None)) and "freezecount" not in key.lower() and "faultcount" not in key.lower():
-        return SensorDeviceClass.TEMPERATURE
-    if unit == "%":
-        return SensorDeviceClass.HUMIDITY
-    if unit == "bar":
-        return SensorDeviceClass.PRESSURE
-    if unit in {"mV", "V"}:
-        return SensorDeviceClass.VOLTAGE
-    if unit == "W":
-        return SensorDeviceClass.POWER
-    # Check if key indicates a timestamp sensor (by suffix or membership in _TIMESTAMP_KEYS)
-    is_timestamp_key = key in _TIMESTAMP_KEYS or any(
-        key.upper().endswith(suffix) for suffix in _TIMESTAMP_SUFFIXES
-    )
-    if is_timestamp_key and key not in _TIME_FORMAT_KEYS:
-        return SensorDeviceClass.TIMESTAMP
-    return None
-
-
-def determine_state_class(key: str) -> SensorStateClass | None:
-    """Determines the appropriate state class for a sensor."""
-    # Check if key indicates a timestamp sensor (by suffix or membership in _TIMESTAMP_KEYS)
-    is_timestamp_key = key in _TIMESTAMP_KEYS or any(
-        key.upper().endswith(suffix) for suffix in _TIMESTAMP_SUFFIXES
-    )
-
-    if key in _ALL_TEXT_SENSORS or is_timestamp_key or key in NO_UNIT_SENSORS:
-        return None
-    # Handle contact sensors (e.g., CLOSE_CONTACT) which return string values like "RELEASED"/"TRIGGERED"
-    if "contact" in key.lower():
-        return None
-    # Count/fault sensors should not have state_class (they are counters, not measurements)
-    if "freezecount" in key.lower() or "faultcount" in key.lower():
-        return None
-    if "total" in key.lower() or "daily" in key.lower():
-        return SensorStateClass.TOTAL_INCREASING
-    return SensorStateClass.MEASUREMENT
-
-
-def get_icon(key: str, unit: str | None, raw_value: Any) -> str:
-    """Determin a sensor."""
-    if key in _BOOLEAN_VALUE_KEYS or _is_boolean_value(raw_value):
-        return "mdi:toggle-switch"
-    if key == "pH_value":
-        return "mdi:flask"
-    if "flow" in key.lower():
-        return "mdi:pump"
-    if unit == "°C":
-        return "mdi:thermometer"
-    if unit == "bar":
-        return "mdi:gauge"
-    if unit in {"mV", "V"}:
-        return "mdi:flash"
-    if key in _TIMESTAMP_KEYS:
-        return "mdi:clock"
-    return "mdi:information"
-
-
-def should_skip_sensor(key: str, raw_value: Any) -> bool:
-    """Determines if a sensor should be skipped and not created."""
-    return (
-        raw_value is None
-        or str(raw_value).strip() in ("[]", "{}", "")
-        or key in _NON_TEMPERATURE_ONEWIRE_KEYS
-        or key.startswith("_")
-    )
+_FLOW_RATE_SOURCE_KEYS = {"ADC3_value", "IMP2_value"}
 
 
 async def async_setup_entry(
@@ -910,31 +129,37 @@ def _create_special_sensors(
     config_entry: ConfigEntry,
     config: dict[str, Any],
 ) -> tuple[list[SensorEntity], set[str]]:
-    """Creates specialized sensors like error codes, flow rate, and diagnostics."""
+    """Creates special sensors that are not part of the standard sensor sets."""
     sensors: list[SensorEntity] = []
     handled_keys: set[str] = set()
 
-    # ✅ DIAGNOSTIC SENSORS: Always create system monitoring sensors
-    sensors.append(VioletSystemHealthSensor(coordinator, config_entry))
-    sensors.append(VioletConnectionLatencySensor(coordinator, config_entry))
-    sensors.append(VioletLastEventAgeSensor(coordinator, config_entry))
-    handled_keys.update({"system_health", "connection_latency", "last_event_age"})
-    _LOGGER.debug(
-        "Diagnostic sensors created (System Health, Connection Latency, Last Event Age)"
+    # System Health & Monitoring Sensors
+    sensors.extend(
+        [
+            VioletSystemHealthSensor(coordinator, config_entry),
+            VioletConnectionLatencySensor(coordinator, config_entry),
+            VioletLastEventAgeSensor(coordinator, config_entry),
+            VioletAPIRequestRateSensor(coordinator, config_entry),
+            VioletAverageLatencySensor(coordinator, config_entry),
+            VioletRecoverySuccessRateSensor(coordinator, config_entry),
+        ]
     )
-
-    # ✅ DIAGNOSTIC SENSORS: Advanced monitoring sensors
-    sensors.append(VioletAPIRequestRateSensor(coordinator, config_entry))
-    sensors.append(VioletAverageLatencySensor(coordinator, config_entry))
-    sensors.append(VioletRecoverySuccessRateSensor(coordinator, config_entry))
     handled_keys.update(
-        {"api_request_rate", "average_latency", "recovery_success_rate"}
+        {
+            "system_health",
+            "connection_latency",
+            "last_event_age",
+            "api_request_rate",
+            "average_latency",
+            "recovery_success_rate",
+        }
     )
     _LOGGER.debug(
-        "Advanced diagnostic sensors created "
-        "(API Request Rate, Average Latency, Recovery Success Rate)"
+        "Diagnostic sensors created (System Health, Connection Latency, Last Event Age, "
+        "API Request Rate, Average Latency, Recovery Success Rate)"
     )
 
+    # Error Code Sensors
     for key in _ERROR_CODE_KEYS:
         if key in coordinator.data and (
             config["create_all"] or key in config["selected_sensors"]
@@ -943,6 +168,7 @@ def _create_special_sensors(
             handled_keys.add(key)
             _LOGGER.debug("Error code sensor created for %s", key)
 
+    # Flow Rate Sensor
     flow_keys_present = any(key in coordinator.data for key in _FLOW_RATE_SOURCE_KEYS)
     flow_selected = (
         config["create_all"] or "flow_rate_adc3_priority" in config["selected_sensors"]
@@ -976,7 +202,7 @@ def _create_special_sensors(
             handled_keys.add(key)
             _LOGGER.debug("Dosing state sensor created for %s", key)
 
-    # Composite State Sensors (PUMPSTATE, HEATERSTATE, SOLARSTATE)
+    # Composite State Sensors (PUMPSTATE, HEATERSTATE, SOLARSTATE, etc.)
     for key, sensor_config in COMPOSITE_STATE_SENSORS.items():
         if key in coordinator.data:
             # Check if feature is enabled
@@ -1041,62 +267,3 @@ def _create_standard_sensors(
         sensors.append(SensorClass(coordinator, config_entry, description))
 
     return sensors
-
-
-def _build_sensor_description(
-    key: str, raw_value: Any, predefined: dict[str, Any]
-) -> SensorEntityDescription:
-    """Builds a SensorEntityDescription for a given sensor key."""
-    predefined_info = predefined.get(key)
-    # Prefer translation key logic if implemented fully,
-    # but for dynamic sensors derived from API keys, we need dynamic names.
-    # We stick to name here but could use key as translation key
-    # if we updated strings.json
-    name = predefined_info["name"] if predefined_info else key.replace("_", " ").title()
-    icon = predefined_info.get("icon") if predefined_info else None
-
-    unit = UNIT_MAP.get(key) if key not in NO_UNIT_SENSORS else None
-    if _is_boolean_value(raw_value):
-        unit = None  # Booleans should not have a unit
-
-    # Force no unit for count/fault sensors (they are counters, not measurements)
-    if "freezecount" in key.lower() or "faultcount" in key.lower():
-        unit = None
-
-    if unit is None and key not in NO_UNIT_SENSORS and not _is_boolean_value(raw_value):
-        for suffix in [
-            "_min",
-            "_max",
-            "_setpoint",
-            "_target",
-            "_faultcount",
-            "_freezecount",
-        ]:
-            if key.endswith(suffix):
-                base_key = key[: -len(suffix)]
-                # IMPORTANT: Don't inherit unit for count/fault sensors
-                if suffix not in ["_faultcount", "_freezecount"]:
-                    if UNIT_MAP.get(base_key):
-                        unit = UNIT_MAP[base_key]
-                        break
-        # Default temperature unit for onewire/temp sensors (but not for counters!)
-        if unit is None and ("temp" in key.lower() or "onewire" in key.lower()) and "freezecount" not in key.lower() and "faultcount" not in key.lower():
-            unit = "°C"
-
-    # Determine state class
-    state_class = determine_state_class(key)
-    # Force state_class to None for contact sensors (they return string values)
-    if "contact" in key.lower():
-        state_class = None
-
-    return SensorEntityDescription(
-        key=key,
-        name=name,
-        icon=icon or get_icon(key, unit, raw_value),
-        native_unit_of_measurement=unit,
-        device_class=determine_device_class(key, unit, raw_value),
-        state_class=state_class,
-        entity_category=EntityCategory.DIAGNOSTIC
-        if key.startswith("SYSTEM_")
-        else None,
-    )
